+++ src/replays.c
@@ -130,9 +130,9 @@ s32 func_800051C4(void) {
 
     if (sReplayGhostBufferSize != 0) {
         // func_80040174 in mio0_decode.s
-        func_80040174((void*) sReplayGhostBuffer, (sReplayGhostBufferSize * 4) + 0x20, (s32) sReplayGhostEncoded);
+        func_80040174((void*) sReplayGhostBuffer, (sReplayGhostBufferSize * 4) + 0x20, (s32) (uintptr_t) sReplayGhostEncoded);
         phi_v0 =
-            mio0encode((s32) sReplayGhostEncoded, (sReplayGhostBufferSize * 4) + 0x20, (s32) gReplayGhostCompressed);
+            mio0encode((s32) (uintptr_t) sReplayGhostEncoded, (sReplayGhostBufferSize * 4) + 0x20, (s32) (uintptr_t) gReplayGhostCompressed);
         return phi_v0 + 0x1e;
     }
     // Empty buffer used to fall off the end; report zero encoded size (the
@@ -450,7 +450,7 @@ void save_player_replay(void) {
     it says "these inputs were played for __ frames" */
     prevInputs = prevInputsWCounter & REPLAY_CLEAR_FRAME_COUNTER;
     // first frame of inputs
-    if ((*sPlayerInputs) == -1) {
+    if ((*sPlayerInputs) == (u32) -1) {
 
         sPlayerInputs[sPlayerInputIdx] = inputs;
